add btmonitor
