+++ app/view/index/download.html
@@ -50,7 +50,7 @@
 						<div class="cont">
 							<div class="top">
 								<div class="title">Windows面板 {:config_get('new_version_win')}</div>
-								<div class="desc">支持Windows Server 2008 R2/2012/2016/2019，64位系统</div>
+								<div class="desc">支持Windows Server 2008 R2/2012/2016/2019/2022，64位系统</div>
 								<div class="mark">
 									<span>操作简单</span>
 									<span>使用方便</span>
@@ -62,6 +62,22 @@
 						</div>
 					</div>
 				</div>
+				{if config_get('new_version_btm')}<div class="disflex flex_lrcenter mt_30 install-list">
+					<div class="install-box monitor">
+						<div class="img">
+							<img src="/static/images/bt_monitor.png" />
+						</div>
+						<div class="cont">
+							<div class="top">
+								<div class="title">堡塔云监控</div>
+								<div class="desc">多服务监控和异常告警通知</div>
+							</div>
+							<div class="bottom">
+								<a class="btn" href="javascript:;" id="goInstallMonitor">查看安装脚本</a>
+							</div>
+						</div>
+					</div>
+				</div>{/if}
 			</div>
 		</div>
 	</div>
@@ -105,7 +121,7 @@
 			</div>
 		</div>
 	</div>
-	<div class="d4" id="instal-windows">
+	<div class="d4" id="instal-windows" style="background-color: #edf6ef;">
 		<div class="wrap">
 			<div class="wrap-title">
 				<div class="text">Windows面板{:config_get('new_version_win')}安装方法</div>
@@ -121,10 +137,40 @@
 				</div>
 			</div>
 			<div class="tips" style="color: orangered; font-weight: 700">
-				<p>注意：仅支持Windows Server 2008 R2/2012/2016/2019，64位系统（中文简体），且未安装其它环境</p>
+				<p>注意：仅支持Windows Server 2008 R2/2012/2016/2019/2022，64位系统（中文简体），且未安装其它环境</p>
 			</div>
 		</div>
 	</div>
+	{if config_get('new_version_btm')}
+	<div class="d4" id="instal-monitor">
+		<div class="wrap">
+			<div class="wrap-title">
+				<div class="text" style="margin-right: 12px;">堡塔云监控{:config_get('new_version_btm')}安装脚本</div>
+			</div>
+			<div class="desc">
+				使用 SSH 连接工具，如
+				<a class="link" href="https://www.putty.org/" target="_blank" rel="noreferrer">PUTTY</a>
+				连接到您的 Linux 服务器后，根据系统执行相应命令开始安装：
+			</div>
+			<div class="install-code">
+				<span class="osname">堡塔云监控安装脚本</span>
+				<div class="code-cont">
+					<div class="command" title="点击复制安装脚本">curl -sS {$siteurl}/install/install_btmonitor.sh -o /tmp/install_btmonitor.sh && bash /tmp/install_btmonitor.sh</div>
+					<span class="ico-copy" title="点击复制安装脚本" data-clipboard-text="curl -sS {$siteurl}/install/install_btmonitor.sh -o /tmp/install_btmonitor.sh && bash /tmp/install_btmonitor.sh">复制</span>
+				</div>
+			</div>
+			<div class="install-code">
+				<span class="osname">堡塔云监控更新脚本</span>
+				<div class="code-cont">
+					<div class="command" title="点击复制更新脚本">curl {$siteurl}/install/update_btmonitor.sh|bash</div>
+					<span class="ico-copy" title="点击复制更新脚本" data-clipboard-text="curl {$siteurl}/install/update_btmonitor.sh|bash">复制</span>
+				</div>
+			</div>
+			<div class="tips" style="color: orangered; font-weight: 700">
+				<p>注意：推荐使用Chrome、火狐、edge浏览器，国产浏览器（极速模式）</p>
+			</div>
+		</div>
+	</div>{/if}
 	<div class="animate-bg"></div>
 </div>
  
@@ -140,7 +186,7 @@
 					</div>
 					<div class="fb2 mt_15">
 						<p>
-							Copyright © 2022  宝塔面板安装脚本
+							Copyright © {:date('Y')}  宝塔面板安装脚本
 						</p>
 					</div>
 				</div>
@@ -184,6 +230,10 @@
 
 		$('#goInstallCloud').click(function () {
 			scrollTop('#instal-cloud');
+		});
+
+		$('#goInstallMonitor').click(function () {
+			scrollTop('#instal-monitor');
 		});
 
 		function GetRequest() {
